IO fixes.
git-svn-id: https://ravl.svn.sourceforge.net/svnroot/ravl/main/Ravl@1873 7226f50c-d813-0410-99bf-979b288ab488

+++ Image/Processing/Filters/HistogramEqualise.hh
@@ -80,7 +80,7 @@ namespace RavlImageN {
   template<class DataT>
   ostream &operator<<(ostream &s,const HistogramEqualiseC<DataT> &hist) {
     int v = 0; // stream version no.
-    s << v << ' ' << hist.MinValue() << ' ' << hist.MaxValue() << ' ' << hist.IsRescale();
+    s << v << ' ' << hist.MinValue() << ' ' << hist.MaxValue() << ' ' << ((int) hist.IsRescale());
     return s;
   }
   //: Write to a stream.
@@ -102,7 +102,7 @@ namespace RavlImageN {
   template<class DataT>
   BinOStreamC &operator<<(BinOStreamC &s,const HistogramEqualiseC<DataT> &hist) {
     int v = 0; // stream version no.
-    s << v << hist.MinValue() << hist.MaxValue() << hist.IsRescale();
+    s << v << hist.MinValue() << hist.MaxValue() << ((int) hist.IsRescale());
     return s;
   }
   //: Write to a binary stream.
